add bulding

+++ 08_networking/include/SFML-Book/common/FuncIds.hpp
@@ -54,7 +54,7 @@ namespace book
         struct Info {
             int cost;
             std::string name;
-            unsigned int makeAs;
+            short int makeAs;
             int icon_id;
             int animation_id;
             std::string description;
